refactor(editor): extract GUI action dispatch into GuiActionHandler
Moves handle_gui_action/2 (66 dispatch clauses + ~30 private helpers)
from MingaEditor into MingaEditor.Handlers.GuiActionHandler, reducing
the Editor GenServer from 3,447 to 2,526 lines.

Shared helpers needed by both the handler and GenServer are promoted
from defp to @doc false def. No behavior changes; all 8,128 tests pass.

Refs #1433
