+++ lib/minga_editor.ex
@@ -17,13 +17,10 @@ defmodule MingaEditor do
   alias MingaEditor.Agent.Events
   alias MingaEditor.Agent.UIState
   alias Minga.Buffer
-  alias Minga.Clipboard
   alias Minga.Config
   alias Minga.Editing.Completion
   alias Minga.FileRef
-  alias Minga.FileWatcher
   alias Minga.Git
-  alias Minga.LSP.Supervisor, as: LspSupervisor
 
   alias Minga.Diagnostics.Decorations, as: DiagDecorations
   alias Minga.Session
@@ -57,6 +54,7 @@ defmodule MingaEditor do
   alias MingaEditor.Viewport
 
   alias MingaEditor.Handlers.FileEventHandler
+  alias MingaEditor.Handlers.GuiActionHandler
   alias MingaEditor.Handlers.HighlightHandler
   alias MingaEditor.Handlers.SessionHandler
   alias MingaEditor.Handlers.ToolHandler
@@ -67,7 +65,6 @@ defmodule MingaEditor do
   alias Minga.Mode
   alias Minga.Project.FileTree
   # PopupLifecycle alias removed: warnings popup replaced by bottom panel (#825)
-  alias MingaEditor.Frontend.Protocol
 
   @typedoc "Options for starting the editor."
   @type start_opt ::
@@ -88,10 +85,7 @@ defmodule MingaEditor do
           | {:session_dir, String.t()}
           | {:suppress_tool_prompts, boolean()}
 
-  alias MingaAgent.Session, as: AgentSession
-
   alias MingaEditor.State, as: EditorState
-  alias MingaEditor.VimState
   alias MingaEditor.Session.State, as: SessionState
   alias MingaEditor.State.LSP, as: LSPState
   alias MingaEditor.State.Session, as: EditorSessionState
@@ -100,11 +94,8 @@ defmodule MingaEditor do
   alias MingaEditor.State.AgentAccess
   alias MingaEditor.State.Buffers
   alias MingaEditor.State.FileTree, as: FileTreeState
-  alias MingaEditor.State.Search, as: SearchData
-  alias MingaEditor.State.Windows
   alias MingaEditor.State.Tab.Context, as: TabContext
 
-  alias MingaEditor.MinibufferData
   alias MingaEditor.MouseHoverTooltip
   alias MingaEditor.PickerUI
 
@@ -562,7 +553,7 @@ defmodule MingaEditor do
 
   def handle_info({:minga_input, {:gui_action, action}}, state) do
     snapshot = Input.Router.capture_snapshot(state)
-    new_state = handle_gui_action(state, action)
+    new_state = GuiActionHandler.dispatch(state, action)
     new_state = Input.Router.post_action_housekeeping(new_state, snapshot)
     {:noreply, new_state}
   end
@@ -1128,10 +1119,11 @@ defmodule MingaEditor do
 
   defp option_source_matches?(_source, _server), do: false
 
+  @doc false
   @spec push_full_config_state(EditorState.t()) :: :ok
-  defp push_full_config_state(%{port_manager: nil}), do: :ok
+  def push_full_config_state(%{port_manager: nil}), do: :ok
 
-  defp push_full_config_state(%{port_manager: port} = state) do
+  def push_full_config_state(%{port_manager: port} = state) do
     if MingaEditor.Frontend.gui?(state.capabilities) do
       config_state =
         MingaEditor.Frontend.Protocol.GUI.config_state(
@@ -1147,10 +1139,11 @@ defmodule MingaEditor do
     :exit, _ -> :ok
   end
 
+  @doc false
   @spec push_config_state_entry(EditorState.t(), atom(), term()) :: EditorState.t()
-  defp push_config_state_entry(%{port_manager: nil} = state, _name, _value), do: state
+  def push_config_state_entry(%{port_manager: nil} = state, _name, _value), do: state
 
-  defp push_config_state_entry(%{port_manager: port} = state, name, value) do
+  def push_config_state_entry(%{port_manager: port} = state, name, value) do
     if MingaEditor.Frontend.gui?(state.capabilities) and
          MingaEditor.Frontend.Protocol.GUI.settings_option?(name) do
       config_state = MingaEditor.Frontend.Protocol.GUI.config_state_entry(name, value)
@@ -1162,8 +1155,9 @@ defmodule MingaEditor do
     :exit, _ -> state
   end
 
+  @doc false
   @spec apply_runtime_config_option(EditorState.t(), atom(), term()) :: EditorState.t()
-  defp apply_runtime_config_option(state, :theme, theme_name) when is_atom(theme_name) do
+  def apply_runtime_config_option(state, :theme, theme_name) when is_atom(theme_name) do
     case MingaEditor.UI.Theme.get(theme_name) do
       {:ok, theme} ->
         if state.port_manager do
@@ -1183,14 +1177,14 @@ defmodule MingaEditor do
     :exit, _ -> state
   end
 
-  defp apply_runtime_config_option(state, name, _value)
-       when name in [:font_family, :font_size, :font_weight, :font_ligatures] do
+  def apply_runtime_config_option(state, name, _value)
+      when name in [:font_family, :font_size, :font_weight, :font_ligatures] do
     Startup.send_font_config(state)
     state
   end
 
-  defp apply_runtime_config_option(state, name, value)
-       when name in [:line_numbers, :wrap, :tab_width] do
+  def apply_runtime_config_option(state, name, value)
+      when name in [:line_numbers, :wrap, :tab_width] do
     Enum.each(runtime_config_buffers(state), fn buffer ->
       Buffer.set_option(buffer, name, value)
     end)
@@ -1200,13 +1194,13 @@ defmodule MingaEditor do
     |> Layout.invalidate()
   end
 
-  defp apply_runtime_config_option(state, :cursorline, _value) do
+  def apply_runtime_config_option(state, :cursorline, _value) do
     state
     |> EditorState.invalidate_all_windows()
     |> Layout.invalidate()
   end
 
-  defp apply_runtime_config_option(state, _name, _value), do: state
+  def apply_runtime_config_option(state, _name, _value), do: state
 
   @spec runtime_config_buffers(EditorState.t()) :: [pid()]
   defp runtime_config_buffers(state) do
@@ -2309,6 +2303,16 @@ defmodule MingaEditor do
 
   defp buffer_tracked_in_tabs?(_state, _pid), do: false
 
+  @spec tab_buffer_list(MingaEditor.State.Tab.t() | term()) :: [pid()]
+  defp tab_buffer_list(%MingaEditor.State.Tab{context: context}) when is_map(context) do
+    case TabContext.to_workspace_map(context) do
+      %{buffers: %Buffers{list: buffers}} -> Enum.filter(buffers, &is_pid/1)
+      _ -> []
+    end
+  end
+
+  defp tab_buffer_list(_tab), do: []
+
   # Like register_buffer but adds the buffer in the background without
   # switching the active window. Used by ensure_buffer_for_path so agent
   # edits don't yank the user away from their current file.
@@ -2325,717 +2329,31 @@ defmodule MingaEditor do
     log_message(state, "Opened (agent): #{file_path}")
   end
 
+  @doc false
   @spec log_message(state(), String.t()) :: state()
-  defp log_message(state, text), do: MessageLog.log(state, text)
+  def log_message(state, text), do: MessageLog.log(state, text)
 
-  # ── GUI action dispatch ────────────────────────────────────────────────
-  # Semantic commands from SwiftUI chrome. Each action maps to existing
-  # editor operations. These are wired up as SwiftUI views are built;
-  # unimplemented actions log a message and return state unchanged.
-
-  @spec handle_gui_action(state(), Protocol.GUI.gui_action()) :: state()
-
-  # Route gui_actions through the active Shell first. Shells that handle
-  # the action return updated state; unhandled actions fall through to
-  # the Traditional-specific handlers below.
-  defp handle_gui_action(state, :system_will_sleep) do
-    Minga.Log.info(:editor, "System will sleep")
-    state
-  end
-
-  defp handle_gui_action(state, :system_did_wake) do
-    Minga.Log.info(:editor, "System did wake; refreshing files, LSP, and git")
-    FileWatcher.check_all()
-    refresh_current_git_repo()
-    LspSupervisor.restart_all_clients()
-    LspSyncServer.resync_buffers(buffers_for_lsp_resync(state))
-
-    state
-    |> EditorState.invalidate_all_windows()
-    |> Layout.invalidate()
-    |> Renderer.render_or_async()
-  end
-
-  defp handle_gui_action(state, :config_query) do
-    push_full_config_state(state)
-    state
-  end
-
-  defp handle_gui_action(state, {:config_update, name, value}) do
-    if MingaEditor.Frontend.Protocol.GUI.settings_option?(name) do
-      case Minga.Config.Options.set(EditorState.options_server(state), name, value) do
-        {:ok, persisted_value} ->
-          Minga.Config.Options.mark_explicit(EditorState.options_server(state), name)
-          Minga.Config.Writer.persist(name, persisted_value)
-
-          state
-          |> apply_runtime_config_option(name, persisted_value)
-          |> push_config_state_entry(name, persisted_value)
-
-        {:error, reason} ->
-          Minga.Log.warning(:config, "Ignored GUI config update for #{inspect(name)}: #{reason}")
-          state
-      end
-    else
-      Minga.Log.warning(
-        :config,
-        "Ignored GUI config update outside settings panel for #{inspect(name)}"
-      )
-
-      state
-    end
-  end
-
-  defp handle_gui_action(%{shell: MingaEditor.Shell.Board} = state, action) do
-    {shell_state, workspace} =
-      MingaEditor.Shell.Board.handle_gui_action(state.shell_state, state.workspace, action)
-
-    state = %{state | shell_state: shell_state, workspace: workspace}
-
-    # After Board zoom into an agent card, atomically activate the
-    # agent view (session, scope, window content, prompt focus).
-    # The Board handler can't do this because it only has
-    # (shell_state, workspace), not the full EditorState.
-    case action do
-      {:board_select_card, card_id} ->
-        card = Map.get(shell_state.cards, card_id)
-
-        {new_board, state} =
-          MingaEditor.Shell.Board.SessionLifecycle.ensure_session(state.shell_state, card, state)
-
-        state = EditorState.update_shell_state(state, fn _ -> new_board end)
-        card = new_board.cards[card_id]
-        MingaEditor.AgentActivation.activate_for_card(state, card)
-
-      _ ->
-        state
-    end
-  end
-
-  defp handle_gui_action(state, {:select_tab, id}) do
-    EditorState.switch_tab(state, id)
-  end
-
-  defp handle_gui_action(state, {:tab_copy_path, id}) do
-    copy_tab_path(state, id)
-  end
-
-  defp handle_gui_action(state, :hover_open_action) do
-    accept_hover_open_action(state)
-  end
-
-  defp handle_gui_action(state, {:close_tab, id}) do
-    # Delegate to the shell: Traditional switches to the target tab when
-    # needed; Board and tab-bar-less Traditional return unchanged.
-    {shell_state, workspace} =
-      state.shell.handle_gui_action(state.shell_state, state.workspace, {:close_tab, id})
-
-    state = %{state | shell_state: shell_state, workspace: workspace}
-
-    # Only close the buffer when the shell has a tab bar.
-    # EditorState.active_tab/1 returns nil when there are no tabs.
-    if EditorState.active_tab(state) do
-      Commands.BufferManagement.execute(state, :force_quit)
-    else
-      state
-    end
-  end
-
-  defp handle_gui_action(state, {:file_tree_click, index}) do
-    gui_tree_action(state, index, :click)
-  end
-
-  defp handle_gui_action(state, {:file_tree_toggle, index}) do
-    gui_tree_action(state, index, :toggle)
-  end
-
-  defp handle_gui_action(state, {:file_tree_open_in_split, index}) do
-    open_file_tree_entry_in_split(state, index)
-  end
-
-  defp handle_gui_action(state, {:file_tree_new_file, index}) do
-    state = move_tree_cursor(state, index)
-    Commands.FileTree.new_file(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_new_folder, index}) do
-    state = move_tree_cursor(state, index)
-    Commands.FileTree.new_folder(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_edit_confirm, text}) do
-    case state.workspace.file_tree.editing do
-      nil ->
-        state
-
-      %{} ->
-        ft = FileTreeState.update_editing_text(state.workspace.file_tree, text)
-
-        state =
-          EditorState.update_workspace(state, &SessionState.set_file_tree(&1, ft))
-
-        Commands.FileTree.confirm_editing(state)
-    end
-  end
-
-  defp handle_gui_action(state, :file_tree_edit_cancel) do
-    Commands.FileTree.cancel_editing(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_delete, index}) do
-    state = move_tree_cursor(state, index)
-    Commands.FileTree.delete(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_rename, index}) do
-    state = move_tree_cursor(state, index)
-    Commands.FileTree.rename(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_duplicate, index}) do
-    state = move_tree_cursor(state, index)
-    Commands.FileTree.duplicate(state)
-  end
-
-  defp handle_gui_action(state, {:file_tree_move, source_index, target_dir_index}) do
-    Commands.FileTree.move(state, source_index, target_dir_index)
-  end
-
-  defp handle_gui_action(state, {:file_tree_drop, intent}) do
-    Commands.FileTree.drop(state, intent)
-  end
-
-  defp handle_gui_action(state, :file_tree_collapse_all) do
-    Commands.FileTree.collapse_all(state)
-  end
-
-  defp handle_gui_action(state, :file_tree_refresh) do
-    Commands.FileTree.refresh(state)
-  end
-
-  defp handle_gui_action(state, {:completion_select, index}) do
-    case MingaEditor.State.ModalOverlay.completion(state) do
-      %Completion{} = comp ->
-        accept_visible_completion(state, comp, index)
-
-      nil ->
-        state
-    end
-  end
-
-  defp handle_gui_action(state, {:breadcrumb_click, _segment_index}) do
-    # Breadcrumb navigation is a follow-up feature.
-    state
-  end
-
-  defp handle_gui_action(state, {:toggle_panel, 0}) do
-    Commands.FileTree.toggle(state)
-  end
-
-  defp handle_gui_action(state, {:toggle_panel, 1}) do
-    EditorState.set_bottom_panel(state, BottomPanel.toggle(EditorState.bottom_panel(state)))
-  end
-
-  defp handle_gui_action(state, {:toggle_panel, 2}) do
-    Commands.Git.execute(state, :git_status_toggle)
-  end
-
-  defp handle_gui_action(state, {:toggle_panel, 3}) do
-    Commands.Agent.toggle_agent_split(state)
-  end
-
-  defp handle_gui_action(state, {:toggle_panel, _panel}) do
-    state
-  end
-
-  defp handle_gui_action(state, :new_tab) do
-    Commands.BufferManagement.execute(state, :new_buffer)
-  end
-
-  defp handle_gui_action(state, {:panel_switch_tab, tab_index}) do
-    EditorState.set_bottom_panel(
-      state,
-      BottomPanel.switch_tab(EditorState.bottom_panel(state), tab_index)
-    )
-  end
-
-  defp handle_gui_action(state, :panel_dismiss) do
-    EditorState.set_bottom_panel(state, BottomPanel.dismiss(EditorState.bottom_panel(state)))
-  end
-
-  defp handle_gui_action(state, {:panel_resize, height_percent}) do
-    EditorState.set_bottom_panel(
-      state,
-      BottomPanel.resize(EditorState.bottom_panel(state), height_percent)
-    )
-  end
-
-  defp handle_gui_action(state, {:open_file, path}) do
-    if File.dir?(path) do
-      open_dropped_directory(state, path)
-    else
-      open_file_by_path(state, path)
-    end
-  end
-
-  defp handle_gui_action(state, {:tool_install, name_str}) do
-    name = String.to_existing_atom(name_str)
-
-    case Minga.Tool.Manager.install(name) do
-      :ok -> EditorState.set_status(state, "Installing #{name_str}...")
-      {:error, reason} -> EditorState.set_status(state, "Cannot install #{name_str}: #{reason}")
-    end
-  rescue
-    ArgumentError -> EditorState.set_status(state, "Unknown tool: #{name_str}")
-  end
-
-  defp handle_gui_action(state, {:tool_uninstall, name_str}) do
-    name = String.to_existing_atom(name_str)
-
-    case Minga.Tool.Manager.uninstall(name) do
-      :ok -> EditorState.set_status(state, "Uninstalled #{name_str}")
-      {:error, reason} -> EditorState.set_status(state, "Cannot uninstall #{name_str}: #{reason}")
-    end
-  rescue
-    ArgumentError -> EditorState.set_status(state, "Unknown tool: #{name_str}")
-  end
-
-  defp handle_gui_action(state, {:tool_update, name_str}) do
-    name = String.to_existing_atom(name_str)
-
-    case Minga.Tool.Manager.update(name) do
-      :ok -> EditorState.set_status(state, "Updating #{name_str}...")
-      {:error, reason} -> EditorState.set_status(state, "Cannot update #{name_str}: #{reason}")
-    end
-  rescue
-    ArgumentError -> EditorState.set_status(state, "Unknown tool: #{name_str}")
-  end
-
-  defp handle_gui_action(state, :tool_dismiss) do
-    # The tool manager panel is closed; no state change needed since
-    # visibility is driven by the BEAM's render cycle
-    state
-  end
-
-  defp handle_gui_action(state, {:agent_tool_toggle, message_index}) do
-    session = AgentAccess.session(state)
-
-    if session do
-      try do
-        AgentSession.toggle_tool_collapse(session, message_index)
-      catch
-        :exit, _ -> :ok
-      end
-    end
-
-    state
-  end
-
-  defp handle_gui_action(state, {:minibuffer_select, index}) do
-    case state.workspace.editing do
-      %{mode: :command, mode_state: ms} ->
-        {candidates, _total} = MinibufferData.complete_ex_command(ms.input)
-        clamped = MinibufferData.clamp_index(index, length(candidates))
-
-        case Enum.at(candidates, clamped) do
-          nil ->
-            state
-
-          %{label: label} ->
-            new_ms = %{ms | input: label, candidate_index: 0}
-            set_vim_mode_state(state, new_ms)
-        end
-
-      _ ->
-        state
-    end
-  end
-
-  defp handle_gui_action(state, {:execute_command, name_str}) do
-    command = String.to_existing_atom(name_str)
-
-    # Discard any follow-up action (dot_repeat, replay_macro): GUI chrome
-    # buttons are not vim editing operations and don't participate in the
-    # action pipeline.
-    case Commands.execute(state, command) do
-      {new_state, _action} -> new_state
-      new_state -> new_state
-    end
-  rescue
-    ArgumentError ->
-      Minga.Log.warning(:editor, "[execute_command] unrecognized command: #{name_str}")
-      state
-  end
-
-  defp handle_gui_action(state, {:git_stage_file, path}) do
-    git_action(
-      state,
-      fn git_root -> Minga.Git.stage(git_root, git_relative_path(git_root, path)) end,
-      "Staged #{path}"
-    )
-  end
-
-  defp handle_gui_action(state, {:git_unstage_file, path}) do
-    git_action(
-      state,
-      fn git_root -> Minga.Git.unstage(git_root, git_relative_path(git_root, path)) end,
-      "Unstaged #{path}"
-    )
-  end
-
-  defp handle_gui_action(state, {:git_discard_file, path}) do
-    git_action(
-      state,
-      fn git_root -> Minga.Git.discard(git_root, git_relative_path(git_root, path)) end,
-      "Discarded #{path}"
-    )
-  end
-
-  defp handle_gui_action(state, :git_stage_all) do
-    git_action(state, fn git_root -> Minga.Git.stage(git_root, ".") end, "Staged all changes")
-  end
-
-  defp handle_gui_action(state, :git_unstage_all) do
-    git_action(state, fn git_root -> Minga.Git.unstage_all(git_root) end, "Unstaged all")
-  end
-
-  defp handle_gui_action(state, {:git_commit, message}) do
-    commit_from_gui(state, message, false)
-  end
-
-  defp handle_gui_action(state, {:git_commit, message, amend?}) do
-    commit_from_gui(state, message, amend?)
-  end
-
-  defp handle_gui_action(state, :git_push) do
-    Commands.Git.execute(state, :git_push)
-  end
-
-  defp handle_gui_action(state, :git_pull) do
-    Commands.Git.execute(state, :git_pull)
-  end
-
-  defp handle_gui_action(state, :git_fetch) do
-    Commands.Git.execute(state, :git_fetch)
-  end
-
-  defp handle_gui_action(state, {:git_commit_amend, message}) do
-    commit_from_gui(state, message, true)
-  end
-
-  defp handle_gui_action(state, {:workspace_close, _ws_id} = action) do
-    {shell_state, workspace} =
-      state.shell.handle_gui_action(state.shell_state, state.workspace, action)
-
-    state
-    |> EditorState.update_shell_state(fn _shell_state -> shell_state end)
-    |> EditorState.update_workspace(fn _workspace -> workspace end)
-    |> EditorState.sync_agent_ui_from_active_workspace()
-  end
-
-  defp handle_gui_action(state, {:workspace_rename, _ws_id, _name} = action) do
-    {shell_state, workspace} =
-      state.shell.handle_gui_action(state.shell_state, state.workspace, action)
-
-    %{state | shell_state: shell_state, workspace: workspace}
-  end
-
-  defp handle_gui_action(state, {:workspace_set_icon, _ws_id, _icon} = action) do
-    {shell_state, workspace} =
-      state.shell.handle_gui_action(state.shell_state, state.workspace, action)
-
-    %{state | shell_state: shell_state, workspace: workspace}
-  end
-
-  defp handle_gui_action(state, {:space_leader_chord, codepoint, modifiers}) do
-    MingaEditor.Input.CUA.SpaceLeader.handle_chord(state, codepoint, modifiers)
-  end
-
-  defp handle_gui_action(state, {:space_leader_retract, codepoint, modifiers}) do
-    MingaEditor.Input.CUA.SpaceLeader.handle_retract(state, codepoint, modifiers)
-  end
-
-  defp handle_gui_action(
-         %{workspace: %{buffers: %{active: buf}}} = state,
-         {:find_pasteboard_search, text, direction}
-       )
-       when is_pid(buf) do
-    # Set the search pattern and execute search_next/search_prev
-    state =
-      EditorState.update_workspace(state, fn ws ->
-        SessionState.set_search(ws, SearchData.record(ws.search, text, :forward))
-      end)
-
-    cmd = if direction == 1, do: :search_prev, else: :search_next
-    MingaEditor.Commands.execute(state, cmd)
-  end
-
-  defp handle_gui_action(state, {:scroll_to_line, line}) do
-    # Scroll the active window's viewport to the target line.
-    active_win_id = state.workspace.windows.active
-    win_map = state.workspace.windows.map
-
-    case Map.get(win_map, active_win_id) do
-      nil ->
-        state
-
-      window ->
-        vp = window.viewport
-        new_vp = Viewport.put_top(vp, max(line, 0))
-        new_win = MingaEditor.Window.set_viewport(window, new_vp)
-        new_map = Map.put(win_map, active_win_id, new_win)
-
-        new_state =
-          EditorState.update_workspace(state, fn ws ->
-            SessionState.set_windows(ws, Windows.set_map(ws.windows, new_map))
-          end)
-
-        Renderer.render_or_async(new_state)
-    end
-  end
-
-  defp handle_gui_action(state, {:fold_toggle_at_line, window_id, line}) do
-    Commands.Folding.execute_at_line(state, window_id, line)
-  end
-
-  defp handle_gui_action(state, :cmd_copy) do
-    Commands.Editing.execute(state, :cmd_copy)
-  end
-
-  defp handle_gui_action(state, :cmd_cut) do
-    Commands.Editing.execute(state, :cmd_cut)
-  end
-
-  defp handle_gui_action(state, {:git_open_file, path}) do
-    case resolve_git_root() do
-      nil ->
-        EditorState.set_status(state, "Not in a git repository")
-
-      git_root ->
-        abs_path = Path.join(git_root, path)
-        open_file_by_path(state, abs_path)
-    end
-  end
-
-  defp handle_gui_action(state, {:git_open_diff, path, section}) do
-    case resolve_git_root() do
-      nil ->
-        EditorState.set_status(state, "Not in a git repository")
-
-      git_root ->
-        open_git_diff_from_panel(state, git_root, path, section)
-    end
-  end
-
-  defp handle_gui_action(state, :git_pull_and_retry) do
-    state
-    |> EditorState.clear_git_toast()
-    |> Commands.Git.execute(:git_pull_and_retry)
-  end
-
-  @spec commit_from_gui(state(), String.t(), boolean()) :: state()
-  defp commit_from_gui(state, message, amend?) do
-    case resolve_git_root() do
-      nil ->
-        EditorState.set_status(state, "Not in a git repository")
-
-      git_root ->
-        opts = if amend?, do: [amend: true], else: []
-        result = Minga.Git.commit(git_root, message, opts)
-        refresh_git_repo(git_root)
-        commit_status(state, result, amend?)
-    end
-  end
-
-  @spec commit_status(state(), {:ok, String.t()} | {:error, String.t()}, boolean()) :: state()
-  defp commit_status(state, {:ok, hash}, true),
-    do: EditorState.set_status(state, "Amended #{hash}")
-
-  defp commit_status(state, {:ok, hash}, false),
-    do: EditorState.set_status(state, "Committed #{hash}")
-
-  defp commit_status(state, {:error, reason}, true),
-    do: EditorState.set_status(state, "Amend failed: #{reason}")
-
-  defp commit_status(state, {:error, reason}, false),
-    do: EditorState.set_status(state, "Commit failed: #{reason}")
-
-  @spec open_git_diff_from_panel(state(), String.t(), String.t(), non_neg_integer()) :: state()
-  defp open_git_diff_from_panel(state, git_root, path, section) do
-    entries = git_status_panel_entries(state)
-
-    matches =
-      case section do
-        section when section in 0..3 ->
-          Enum.filter(entries, &git_status_entry_matches?(&1, path, section))
-
-        _legacy ->
-          Enum.filter(entries, &(&1.path == path))
-      end
-
-    case matches do
-      [%Git.StatusEntry{} = entry] ->
-        open_git_diff_for_entry(state, git_root, entry)
-
-      [] ->
-        EditorState.set_status(state, "No git diff entry for #{path}")
-
-      [_ | [_ | _]] ->
-        EditorState.set_status(
-          state,
-          "Ambiguous git diff entry for #{path}; use section-aware diff"
-        )
-    end
-  end
-
-  @spec git_status_entry_matches?(Git.StatusEntry.t(), String.t(), non_neg_integer()) :: boolean()
-  defp git_status_entry_matches?(%Git.StatusEntry{} = entry, path, section) do
-    entry.path == path && git_status_section(entry) == section
-  end
-
-  @spec git_status_section(Git.StatusEntry.t()) :: non_neg_integer()
-  defp git_status_section(%Git.StatusEntry{staged: true}), do: 0
-  defp git_status_section(%Git.StatusEntry{status: :untracked}), do: 2
-  defp git_status_section(%Git.StatusEntry{status: :conflict}), do: 3
-  defp git_status_section(%Git.StatusEntry{}), do: 1
-
-  @spec git_status_panel_entries(state()) :: [Git.StatusEntry.t()]
-  defp git_status_panel_entries(state) do
-    case EditorState.git_status_panel(state) do
-      nil -> []
-      panel -> Map.get(panel, :entries) || []
-    end
-  end
-
-  @spec open_git_diff_for_entry(state(), String.t(), Git.StatusEntry.t()) :: state()
-  defp open_git_diff_for_entry(state, git_root, %Git.StatusEntry{} = entry) do
-    abs_path = git_status_abs_path(git_root, entry.path)
-    git_path = Path.relative_to(abs_path, git_root)
-    git_entry = %{entry | path: git_path}
-
-    case git_diff_content(git_root, abs_path, git_entry) do
-      {:ok, current_content} ->
-        Commands.Git.open_diff_for_path(state, git_root, git_path, abs_path, current_content,
-          staged: entry.staged
-        )
-
-      {:error, message} ->
-        EditorState.set_status(state, message)
-    end
-  end
-
-  @spec git_diff_content(String.t(), String.t(), Git.StatusEntry.t()) ::
-          {:ok, String.t()} | {:error, String.t()}
-  defp git_diff_content(_git_root, _abs_path, %Git.StatusEntry{status: :deleted}) do
-    {:ok, ""}
-  end
-
-  defp git_diff_content(git_root, _abs_path, %Git.StatusEntry{path: rel_path, staged: true}) do
-    case Git.show_staged(git_root, rel_path) do
-      {:ok, content} -> {:ok, content}
-      :error -> {:error, "Could not read staged file: #{rel_path}"}
-    end
-  end
-
-  defp git_diff_content(_git_root, abs_path, %Git.StatusEntry{}) do
-    case File.read(abs_path) do
-      {:ok, current_content} -> {:ok, current_content}
-      {:error, reason} -> {:error, "Could not read file: #{inspect(reason)}"}
-    end
-  end
-
-  @spec git_relative_path(String.t(), String.t()) :: String.t()
-  defp git_relative_path(git_root, path) do
-    git_root
-    |> git_status_abs_path(path)
-    |> Path.relative_to(git_root)
-  end
-
-  @spec git_status_abs_path(String.t(), String.t()) :: String.t()
-  defp git_status_abs_path(git_root, path) do
-    project_root = Minga.Project.resolve_root()
-
-    if String.starts_with?(Path.expand(project_root), Path.expand(git_root)) do
-      Path.join(project_root, path)
-    else
-      Path.join(git_root, path)
-    end
-  end
-
-  @spec accept_visible_completion(state(), Completion.t(), non_neg_integer()) :: state()
-  defp accept_visible_completion(state, comp, index) do
-    {visible, _selected_offset} = Completion.visible_items(comp)
-
-    case Enum.at(visible, index) do
-      nil ->
-        state
-
-      _item ->
-        updated = Completion.select_visible(comp, index)
-
-        do_accept_completion(
-          MingaEditor.State.ModalOverlay.update_completion(state, fn _ -> updated end),
-          updated
-        )
-    end
-  end
-
-  # Project.switch/1 is a cast; the picker opens against current state while the
-  # file cache rebuilds asynchronously. BEAM message ordering from the Editor
-  # process guarantees the cast reaches Project before any subsequent call.
-  @spec open_dropped_directory(state(), String.t()) :: state()
-  defp open_dropped_directory(state, dir_path) do
-    Minga.Project.switch(dir_path)
-    PickerUI.open(state, MingaEditor.UI.Picker.FileSource)
-  end
-
-  # Moves the tree cursor to a specific index (used by GUI context menu / header actions).
-  @spec move_tree_cursor(state(), non_neg_integer()) :: state()
-  defp move_tree_cursor(%{workspace: %{file_tree: %{tree: nil}}} = state, _index), do: state
-
-  defp move_tree_cursor(state, index) do
-    EditorState.update_workspace(state, fn ws ->
-      tree = FileTree.select(ws.file_tree.tree, index)
-      SessionState.set_file_tree(ws, FileTreeState.set_tree(ws.file_tree, tree))
-    end)
-  end
-
-  @spec git_action(state(), (String.t() -> :ok | {:error, String.t()}), String.t()) :: state()
-  defp git_action(state, operation, success_msg) when is_binary(success_msg) do
-    git_root = resolve_git_root()
-
-    if git_root do
-      case operation.(git_root) do
-        :ok ->
-          refresh_git_repo(git_root)
-          EditorState.set_status(state, success_msg)
-
-        {:error, reason} ->
-          EditorState.set_status(state, "Git error: #{reason}")
-      end
-    else
-      EditorState.set_status(state, "Not in a git repository")
-    end
-  end
-
+  @doc false
   @spec open_file_by_path(state(), String.t()) :: state()
-  defp open_file_by_path(state, abs_path) do
+  def open_file_by_path(state, abs_path) do
     case open_file_by_path_result(state, abs_path) do
       {:ok, new_state} -> new_state
       {:error, _reason} -> EditorState.set_status(state, "Could not open #{abs_path}")
     end
   end
 
+  @doc false
   @spec open_file_by_path_result(state(), String.t()) :: {:ok, state()} | {:error, term()}
-  defp open_file_by_path_result(state, abs_path) do
+  def open_file_by_path_result(state, abs_path) do
     case file_tab_for_path_in_active_workspace(state, abs_path) do
       %Tab{id: id} -> {:ok, EditorState.switch_tab(state, id)}
       nil -> start_and_register_file(state, abs_path)
     end
   end
 
+  @doc false
   @spec start_and_register_file(state(), String.t()) :: {:ok, state()} | {:error, term()}
-  defp start_and_register_file(state, abs_path) do
+  def start_and_register_file(state, abs_path) do
     case Commands.start_buffer(abs_path, EditorState.options_server(state)) do
       {:ok, pid} ->
         new_state = register_buffer(state, pid, abs_path)
@@ -3046,56 +2364,12 @@ defmodule MingaEditor do
     end
   end
 
-  @spec open_file_by_path_in_active_window(state(), String.t()) :: state()
-  defp open_file_by_path_in_active_window(state, abs_path) do
-    case file_tab_for_path_in_active_workspace(state, abs_path) do
-      %Tab{} = tab ->
-        open_tab_buffer_in_active_window(state, tab, abs_path)
-
-      nil ->
-        case Commands.start_buffer(abs_path, EditorState.options_server(state)) do
-          {:ok, pid} -> register_buffer_in_active_window(state, pid, abs_path)
-          {:error, _reason} -> EditorState.set_status(state, "Could not open #{abs_path}")
-        end
-    end
-  end
-
-  @spec open_tab_buffer_in_active_window(state(), Tab.t(), String.t()) :: state()
-  defp open_tab_buffer_in_active_window(state, tab, abs_path) do
-    case tab_active_buffer(tab) do
-      pid when is_pid(pid) -> show_buffer_in_active_window(state, pid)
-      nil -> EditorState.set_status(state, "Could not open #{abs_path}")
-    end
-  end
-
-  @spec show_buffer_in_active_window(state(), pid()) :: state()
-  defp show_buffer_in_active_window(state, pid) when is_pid(pid) do
-    EditorState.update_workspace(state, fn workspace ->
-      buffers =
-        case Enum.find_index(workspace.buffers.list, &(&1 == pid)) do
-          nil -> Buffers.add(workspace.buffers, pid)
-          idx -> Buffers.switch_to(workspace.buffers, idx)
-        end
-
-      workspace
-      |> SessionState.set_buffers(buffers)
-      |> SessionState.sync_active_window_buffer()
-    end)
-  end
-
-  @spec tab_active_buffer(Tab.t()) :: pid() | nil
-  defp tab_active_buffer(%Tab{context: context}) when is_map(context) do
-    case TabContext.to_workspace_map(context) do
-      %{buffers: %Buffers{active: pid}} when is_pid(pid) -> pid
-      _ -> nil
-    end
-  end
-
+  @doc false
   @spec file_tab_for_path_in_active_workspace(state(), String.t()) :: Tab.t() | nil
-  defp file_tab_for_path_in_active_workspace(
-         %{shell_state: %{tab_bar: %TabBar{} = tb}} = state,
-         path
-       ) do
+  def file_tab_for_path_in_active_workspace(
+        %{shell_state: %{tab_bar: %TabBar{} = tb}} = state,
+        path
+      ) do
     file_ref = FileRef.new(path)
 
     if active_buffer_matches_file_ref?(state, file_ref) do
@@ -3105,7 +2379,7 @@ defmodule MingaEditor do
     end
   end
 
-  defp file_tab_for_path_in_active_workspace(_state, _path), do: nil
+  def file_tab_for_path_in_active_workspace(_state, _path), do: nil
 
   @spec active_buffer_matches_file_ref?(state(), FileRef.t()) :: boolean()
   defp active_buffer_matches_file_ref?(
@@ -3131,36 +2405,9 @@ defmodule MingaEditor do
     :exit, _ -> nil
   end
 
-  @spec register_buffer_in_active_window(state(), pid(), String.t()) :: state()
-  defp register_buffer_in_active_window(state, buffer_pid, file_path) do
-    state =
-      state
-      |> EditorState.update_workspace(fn workspace ->
-        workspace
-        |> SessionState.set_buffers(Buffers.add(workspace.buffers, buffer_pid))
-        |> SessionState.sync_active_window_buffer()
-      end)
-      |> EditorState.monitor_buffer(buffer_pid)
-
-    state = log_message(state, "Opened: #{file_path}")
-
-    Minga.Events.broadcast(
-      :buffer_opened,
-      %Minga.Events.BufferEvent{buffer: buffer_pid, path: file_path},
-      EditorState.events_registry(state)
-    )
-
-    state = HighlightSync.setup_for_buffer_pid(state, buffer_pid)
-
-    if state.backend != :headless do
-      Process.send_after(self(), :request_code_lens_and_inlay_hints, 800)
-    end
-
-    state
-  end
-
+  @doc false
   @spec resolve_git_root() :: String.t() | nil
-  defp resolve_git_root do
+  def resolve_git_root do
     root = Minga.Project.resolve_root()
 
     case Minga.Git.root_for(root) do
@@ -3169,119 +2416,9 @@ defmodule MingaEditor do
     end
   end
 
-  @spec refresh_current_git_repo() :: :ok
-  defp refresh_current_git_repo do
-    case resolve_git_root() do
-      nil -> :ok
-      git_root -> refresh_git_repo(git_root)
-    end
-  end
-
-  @spec accept_hover_open_action(state()) :: state()
-  defp accept_hover_open_action(state) do
-    case EditorState.hover_popup(state) do
-      %MingaEditor.HoverPopup{open_action: action} when action != nil ->
-        state = EditorState.dismiss_hover_popup(state)
-        execute_hover_open_action(state, action)
-
-      _ ->
-        state
-    end
-  end
-
-  @spec execute_hover_open_action(state(), MingaEditor.HoverPopup.open_action()) :: state()
-  defp execute_hover_open_action(state, {:goto_location, uri, line, col}) do
-    LspActions.open_location(state, uri, line, col)
-  end
-
-  defp execute_hover_open_action(state, action) when is_atom(action) do
-    case Commands.execute(state, action) do
-      {new_state, _action} -> new_state
-      new_state -> new_state
-    end
-  end
-
-  @spec copy_tab_path(state(), Tab.id()) :: state()
-  defp copy_tab_path(state, id) do
-    case tab_file_path(state, id) do
-      nil ->
-        EditorState.set_status(state, "Tab has no file path")
-
-      path ->
-        Clipboard.write_async(path)
-        maybe_send_gui_clipboard_write(state, path)
-        EditorState.set_status(state, "Copied #{path}")
-    end
-  end
-
-  @spec maybe_send_gui_clipboard_write(state(), String.t()) :: :ok
-  defp maybe_send_gui_clipboard_write(%{port_manager: nil}, _path), do: :ok
-
-  defp maybe_send_gui_clipboard_write(state, path) do
-    MingaEditor.Frontend.clipboard_write(state.port_manager, path)
-  end
-
-  @spec tab_file_path(state(), Tab.id()) :: String.t() | nil
-  defp tab_file_path(state, id) do
-    case EditorState.tab_bar(state) do
-      %TabBar{} = tb -> tab_file_path_from_tab(state, tb, TabBar.get(tb, id))
-      nil -> nil
-    end
-  end
-
-  @spec tab_file_path_from_tab(state(), TabBar.t(), Tab.t() | nil) :: String.t() | nil
-  defp tab_file_path_from_tab(_state, _tb, nil), do: nil
-  defp tab_file_path_from_tab(_state, _tb, %Tab{kind: :agent}), do: nil
-
-  defp tab_file_path_from_tab(state, %TabBar{active_id: active_id} = tb, %Tab{id: id}) do
-    case id == active_id do
-      true -> active_buffer_path(state)
-      false -> inactive_tab_path(TabBar.get(tb, id))
-    end
-  end
-
-  @spec active_buffer_path(state()) :: String.t() | nil
-  defp active_buffer_path(%{workspace: %{buffers: %{active: buf}}}) when is_pid(buf) do
-    Buffer.file_path(buf)
-  end
-
-  defp active_buffer_path(_state), do: nil
-
-  @spec inactive_tab_path(Tab.t() | nil) :: String.t() | nil
-  defp inactive_tab_path(%Tab{context: context}) when is_map(context) do
-    case TabContext.to_workspace_map(context) do
-      %{buffers: %Buffers{active: pid}} when is_pid(pid) -> Buffer.file_path(pid)
-      _ -> nil
-    end
-  end
-
-  defp inactive_tab_path(_tab), do: nil
-
-  @spec buffers_for_lsp_resync(state()) :: [pid()]
-  defp buffers_for_lsp_resync(state) do
-    active_buffers = Enum.filter(state.workspace.buffers.list, &is_pid/1)
-
-    tab_buffers =
-      case EditorState.tab_bar(state) do
-        %MingaEditor.State.TabBar{tabs: tabs} -> Enum.flat_map(tabs, &tab_buffer_list/1)
-        _ -> []
-      end
-
-    Enum.uniq(active_buffers ++ tab_buffers)
-  end
-
-  @spec tab_buffer_list(MingaEditor.State.Tab.t() | term()) :: [pid()]
-  defp tab_buffer_list(%MingaEditor.State.Tab{context: context}) when is_map(context) do
-    case TabContext.to_workspace_map(context) do
-      %{buffers: %Buffers{list: buffers}} -> Enum.filter(buffers, &is_pid/1)
-      _ -> []
-    end
-  end
-
-  defp tab_buffer_list(_tab), do: []
-
+  @doc false
   @spec refresh_git_repo(String.t()) :: :ok
-  defp refresh_git_repo(git_root) do
+  def refresh_git_repo(git_root) do
     case Git.lookup_repo(git_root) do
       nil -> :ok
       pid -> Git.Repo.refresh(pid)
@@ -3305,55 +2442,6 @@ defmodule MingaEditor do
   defp maybe_refresh_tool_picker(state), do: state
 
   # maybe_show_tool_prompt moved to ToolHandler
-
-  @spec open_file_tree_entry_in_split(state(), non_neg_integer()) :: state()
-  defp open_file_tree_entry_in_split(%{workspace: %{file_tree: %{tree: nil}}} = state, _index) do
-    state
-  end
-
-  defp open_file_tree_entry_in_split(state, index) do
-    state =
-      state
-      |> move_tree_cursor(index)
-      |> unfocus_file_tree_for_split()
-
-    case FileTree.selected_entry(state.workspace.file_tree.tree) do
-      %{dir?: false, path: path} ->
-        state
-        |> Commands.Movement.execute(:split_vertical)
-        |> Commands.Movement.execute(:window_right)
-        |> open_file_by_path_in_active_window(path)
-
-      %{dir?: true} ->
-        state
-
-      nil ->
-        state
-    end
-  end
-
-  @spec unfocus_file_tree_for_split(state()) :: state()
-  defp unfocus_file_tree_for_split(state) do
-    EditorState.update_workspace(state, fn workspace ->
-      workspace
-      |> SessionState.set_file_tree(MingaEditor.State.FileTree.unfocus(workspace.file_tree))
-      |> SessionState.set_keymap_scope(:editor)
-    end)
-  end
-
-  # Moves the file tree cursor to the given index and performs the action.
-  @spec gui_tree_action(state(), non_neg_integer(), :click | :toggle) :: state()
-  defp gui_tree_action(%{workspace: %{file_tree: %{tree: nil}}} = state, _index, _action),
-    do: state
-
-  defp gui_tree_action(state, index, action) do
-    state = move_tree_cursor(state, index)
-
-    case action do
-      :click -> Commands.FileTree.open_or_toggle(state)
-      :toggle -> Commands.FileTree.open_or_toggle(state)
-    end
-  end
 
   # ── Warning popup debounce ───────────────────────────────────────────────
 
@@ -3435,13 +2523,4 @@ defmodule MingaEditor do
   @doc false
   @spec do_dismiss_completion(state()) :: state()
   defdelegate do_dismiss_completion(state), to: CompletionHandling, as: :dismiss
-
-  # Sets the manual workspace label to the project directory name.
-
-  @spec set_vim_mode_state(state(), term()) :: state()
-  defp set_vim_mode_state(state, new_ms) do
-    EditorState.update_workspace(state, fn ws ->
-      SessionState.update_editing(ws, &VimState.set_mode_state(&1, new_ms))
-    end)
-  end
 end
